port importer/exporter updates from SoD repo
