+++ ui/core/components/individual_sim_ui/importers/individual_addon_importer.tsx
@@ -1,0 +1,136 @@
+import { JsonObject } from '@protobuf-ts/runtime';
+import { ref } from 'tsx-vanilla';
+
+import { IndividualSimUI } from '../../../individual_sim_ui';
+import { Class, EquipmentSpec, Glyphs, Profession, Race, Spec } from '../../../proto/common';
+import { Database } from '../../../proto_utils/database';
+import { nameToClass, nameToProfession, nameToRace } from '../../../proto_utils/names';
+import { classGlyphsConfig } from '../../../talents/factory';
+import { GlyphConfig } from '../../../talents/glyphs_picker';
+import Toast from '../../toast';
+import { IndividualImporter } from './individual_importer';
+
+export class IndividualAddonImporter<SpecType extends Spec> extends IndividualImporter<SpecType> {
+	constructor(parent: HTMLElement, simUI: IndividualSimUI<SpecType>) {
+		super(parent, simUI, { title: 'Addon Import', allowFileUpload: true });
+
+		const warningRef = ref<HTMLDivElement>();
+		this.descriptionElem.appendChild(
+			<>
+				<p>
+					Import settings from the{' '}
+					<a href="https://www.curseforge.com/wow/addons/wowsimsexporter" target="_blank">
+						WoWSims Exporter In-Game Addon
+					</a>
+					.
+				</p>
+				<p>
+					This feature imports gear, race, talents, glyphs, and professions. It does NOT import buffs, debuffs, consumes, rotation, or custom stats.
+				</p>
+				<p>To import, paste the output from the addon below and click, 'Import'.</p>
+				<div ref={warningRef} />
+			</>,
+		);
+
+		if (warningRef.value)
+			new Toast({
+				title: 'Reforging issues',
+				body: (
+					<>
+						There are known issues with Reforging when using the WSE addon.
+						<br />
+						Always make sure to double check your reforges after importing.
+					</>
+				),
+				additionalClasses: ['toast-import-warning'],
+				container: warningRef.value,
+				variant: 'warning',
+				canClose: false,
+				autoShow: true,
+				autohide: false,
+			});
+	}
+
+	async onImport(data: string) {
+		let importJson: any | null;
+		try {
+			importJson = JSON.parse(data);
+		} catch {
+			throw new Error('Please use a valid Addon export.');
+		}
+
+		// Parse all the settings.
+		const charClass = nameToClass((importJson['class'] as string) || '');
+		if (charClass == Class.ClassUnknown) {
+			throw new Error('Could not parse Class!');
+		}
+
+		const race = nameToRace((importJson['race'] as string) || '');
+		if (race == Race.RaceUnknown) {
+			throw new Error('Could not parse Race!');
+		}
+
+		const professions = (importJson['professions'] as Array<{ name: string; level: number }>).map(profData => nameToProfession(profData.name));
+		professions.forEach((prof, i) => {
+			if (prof == Profession.ProfessionUnknown) {
+				throw new Error(`Could not parse profession '${importJson['professions'][i]}'`);
+			}
+		});
+
+		const talentsStr = (importJson['talents'] as string) || '';
+		const glyphsConfig = classGlyphsConfig[charClass];
+
+		const db = await Database.get();
+		const primeGlyphIDs = (importJson['glyphs']['prime'] as Array<string | JsonObject>).map(g => glyphToID(g, db, glyphsConfig.primeGlyphs));
+		const majorGlyphIDs = (importJson['glyphs']['major'] as Array<string | JsonObject>).map(g => glyphToID(g, db, glyphsConfig.majorGlyphs));
+		const minorGlyphIDs = (importJson['glyphs']['minor'] as Array<string | JsonObject>).map(g => glyphToID(g, db, glyphsConfig.minorGlyphs));
+
+		const glyphs = Glyphs.create({
+			prime1: primeGlyphIDs[0] || 0,
+			prime2: primeGlyphIDs[1] || 0,
+			prime3: primeGlyphIDs[2] || 0,
+			major1: majorGlyphIDs[0] || 0,
+			major2: majorGlyphIDs[1] || 0,
+			major3: majorGlyphIDs[2] || 0,
+			minor1: minorGlyphIDs[0] || 0,
+			minor2: minorGlyphIDs[1] || 0,
+			minor3: minorGlyphIDs[2] || 0,
+		});
+
+		const gearJson = importJson['gear'];
+		gearJson.items = (gearJson.items as Array<any>).filter(item => item != null);
+		delete gearJson.version;
+
+		(gearJson.items as Array<any>).forEach(item => {
+			if (item.gems) {
+				item.gems = (item.gems as Array<any>).map(gem => gem || 0);
+			}
+		});
+		const equipmentSpec = EquipmentSpec.fromJson(gearJson);
+
+		this.finishIndividualImport(this.simUI, charClass, race, equipmentSpec, talentsStr, glyphs, professions);
+	}
+}
+
+function glyphNameToID(glyphName: string, glyphsConfig: Record<number, GlyphConfig>): number {
+	if (!glyphName) {
+		return 0;
+	}
+
+	for (const glyphIDStr in glyphsConfig) {
+		if (glyphsConfig[glyphIDStr].name == glyphName) {
+			return parseInt(glyphIDStr);
+		}
+	}
+	throw new Error(`Unknown glyph name '${glyphName}'`);
+}
+
+function glyphToID(glyph: string | JsonObject, db: Database, glyphsConfig: Record<number, GlyphConfig>): number {
+	if (typeof glyph === 'string') {
+		// Legacy version: AddOn exports Glyphs by name (string) only. Names must be in English.
+		return glyphNameToID(glyph, glyphsConfig);
+	}
+
+	// Cata version exports glyph information in a table that includes the name and the glyph spell ID.
+	return db.glyphSpellToItemId(glyph.spellID as number);
+}
