+++ ui/core/sim.ts
@@ -54,17 +54,6 @@ export type StatWeightsData = {
 	result: StatWeightsResult;
 };
 
-export enum SimSettingCategories {
-	Gear = 0,
-	Talents,
-	Rotation,
-	Consumes,
-	Miscellaneous, // Spec-specific settings, Distance from target, tank status, etc
-	External, // Buffs and debuffs
-	Encounter,
-	UISettings, // # iterations, EP weights, filters, etc
-}
-
 interface SimProps {
 	// The type of sim. Default `SimType.SimTypeIndividual`
 	type?: SimType;
