+++ ui/raid/components/exporters/index.ts
@@ -1,0 +1,1 @@
+export { RaidJsonExporter } from './raid_json_exporter';
