+++ ui/scss/core/components/_exporters.scss
@@ -5,10 +5,13 @@
 		resize: none;
 	}
 
-	.link-exporter-pickers {
+	.exporter-category-pickers {
 		display: grid;
-		grid-template-columns: repeat(3, 1fr);
-		column-gap: 1.25rem;
+		grid-template-columns: repeat(4, 1fr);
+		gap: var(--spacer-1) var(--bs-modal-padding);
+		@include media-breakpoint-down(md) {
+			grid-template-columns: repeat(3, 1fr);
+		}
 	}
 
 	.modal-footer {
