+++ ui/scss/core/components/_importers.scss
@@ -13,6 +13,5 @@
 
 	.toast-import-warning {
 		width: 100%;
-		margin-bottom: var(--spacer-3);
 	}
 }
